+++ Reconciliation/Properties/Resources.Designer.cs
@@ -99,5 +99,25 @@ namespace Reconciliation.Properties {
                 return ((System.Drawing.Bitmap)(obj));
             }
         }
+
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
+        internal static System.Drawing.Bitmap MSPHubLogo {
+            get {
+                object obj = ResourceManager.GetObject("MSPHubLogo", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
+        internal static System.Drawing.Bitmap MicrosoftLogo {
+            get {
+                object obj = ResourceManager.GetObject("MicrosoftLogo", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
     }
 }